Test is written in main.go
In main.go, I initialized mymap structure and filled
it with random 100000 data which is the maximum size of mymap.
Then I calculated the reaching 0 and reachinh 99999 keys efficiency.
Like reaching the first element pushed and last element pushed.
Then I do the same thing for an array of size 100000 and filed with
same key value pairs. Then I searched the key 0 and 99999 both on this
array and calculated their efficiency. As a result I printed the
efficiency of first and last elements pushed to the mymap and normal
array in nanoseconds.